Add mongoDb document mangement with repositoryPattern
Implement classes to manange mongodb documents using repository pattern.

For data persist, we use mongodb. While using onion architecture, it's important to restricts other classes to work directly with the data of the application. So Repository pattern was the right pattern to implement.

He comes with several advantages 
- Tight coupling
- Easy to test
and more...